project-eyeseeyou-main/requirements.txt
# Project dependencies
numpy>=1.19.0
torch>=1.8.0
torchvision>=0.9.0
matplotlib>=3.3.0
pillow>=8.0.0
tqdm>=4.50.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
pyswip>=0.2.10
pyyaml>=5.4.0
flask>=2.0.0
streamlit>=1.0.0
ultralytics>=8.0.0
flask-cors>=3.0.10
